Fixes instructions on importing ANYmal URDF in docs (#1915)
# Description

Previously, the ANYmal URDF importer example was setting the stiffness
and damping to new defaults (100.0 and 0.1 respectively). However, this
is undesirable as we want to import the asset with no stiffness and
damping (as we do effort control). This MR corrects the instructions to
convert ANYmal asset.

## Type of change

- Bug fix (non-breaking change which fixes an issue)
- This change requires a documentation update

## Checklist

- [x] I have run the [`pre-commit` checks](https://pre-commit.com/) with
`./isaaclab.sh --format`
- [x] I have made corresponding changes to the documentation
- [x] My changes generate no new warnings
- [ ] I have added tests that prove my fix is effective or that my
feature works
- [ ] I have updated the changelog and the corresponding version in the
extension's `config/extension.toml` file
- [x] I have added my name to the `CONTRIBUTORS.md` or my name already
exists there