Removes `disable_contact_processing` flag from SimulationContext (#1861)
# Description

The `disable_contact_processing` flag inside the SimulationContext
causes confusion, as we were setting it to True in many environments,
but it should be False since we want PhysX to report the contacts.

This MR changes the default behavior now. It is always disabled (i.e. no
contact reporting), until a contact sensor is created. It makes sure
that the performance isn't hampered in situations where the user doesn't
want to create any contact sensors.

## Type of Change

- Bug fix (non-breaking change which fixes an issue)
- Breaking change (fix or feature that would cause existing
functionality to not work as expected)

## Checklist

- [x] I have run the [`pre-commit` checks](https://pre-commit.com/) with
`./isaaclab.sh --format`
- [x] I have made corresponding changes to the documentation
- [ ] My changes generate no new warnings
- [ ] I have added tests that prove my fix is effective or that my
feature works
- [x] I have updated the changelog and the corresponding version in the
extension's `config/extension.toml` file
- [x] I have added my name to the `CONTRIBUTORS.md` or my name already
exists there